[OPT-421] steer: loop over the recirculation group and send the product IDs to the compare endpoint (#1741)
* initial commit for drawer

* drawer intermission

* added css transition

* intermission

* added css animations

* added backdrop style

* completed drawer

* fixed timeout

* added drawer with header

* compeleted drawer

* added styles for storybook

* changed icons

* added accessiblity

* added ProductToCompare component

* added ListToCompare

* added AddTire and restructure drawer

* created listToCompare

* added useSVGString

* changed add tire

* updated product to compare

* modified compare drawer

* removed unused lines

* added styles for XL

* added buttons

* added position

* changed props styles

* validate isDisabled compare

* initial commit for CompareModal

* changed CompareDrawer

* added backdrop

* changed file name

* removed backdrop

* changed ListToCompare stories

* added CompareModal context

* changed product to compare

* initial commit for CompareTable

* added sub component

* added star component

* changed styles for table

* resolved comment

* removed unused lines

* added rating bar

* added single accordion

* added some component to Comparetable

* fixed types

* added sticky

* intermission

* added column sticky

* added scroll subscription

* added sync table and carousel

* added scrolling

* completed modal

* added close button

* data definition

* fixed table cells

* fixed comparetalbe story

* resolve comments

* added price with promotion

* added animation

* fixed linting errors

* added check button

* added check button to advanced listing

* refactored compare table models

* refactored compare table models2

* fixed issues

* fixed issues

* added components

* changed scroll

* fixed linting errors

* changed styles

* added api

* changed api in pages

* added table factory

* fixed table factory

* intermission for compare utils

* test intermission storybook

* added fetch api

* added remove functionality

* added table cell animation

* fixed styles

* added addToCart

* added use local storage

* added for add to cart

* changed road hazard

* changed hazard modal

* changed styles

* added api

* fixed styles

* changed add to cart

* revert url.ts

* cached useLocalstorage

* removed a file

* changed quantity selector test file

* added groupId

* finalize

* fixed an issue

* fixed a bug

* removed a file

* resolved comment

* resolved comment

* changed styles

* changed image to sidewall type

* changed modal

* changed compare table styles

* changed single accordion

* revert url.ts

* changed time

* fixed a logic for opening drawer

* added slide up animation for drawer

* changed time interval when theme changing

* go mutilple slide per click

* fixed an error when removing

* added classNames for debuggin

* refactored ProductToCompare

* refactored ProductToCompare to TireWithInfoList

* added sticky

* changed loading

* intermission for fix

* changed open flow of quantity selector

* removed right margin

* added css animations

* completed animation

* added styles for header in compare drawer

* changed flow of openging drawer

* changed flow of openging drawer

* fixed the issue

* fixed the v2

* speed up the animatin

* added hover state

* added hover state

* added bodyScrollOption

* changed layout of compare modal -test

* removed body scroll lock

* added animation

* finalize

* intermission for fix

* created new scroll

* added orange  background

* Added max

* test react-sync-scroll

* changed useXSticky

* tested adding two at a time

* updated scrollin exp

* added empty tire

* changed addtoTire

* changed logic to open

* added SyncScroll hooks

* changed styles for compare modal

* changed caption position

* removed unused line

* changed modal width

* fixed final comments

* fixed input height in ios

* changed styles for check label

* made changes for other pages

* changed styles for label

* added api popular compare

* added styles for popular section

* changed useElement

* changed useScrollbar

* added styles for wrappers

* ommit most popular from recirulatin

* added left and right padding styles

* added add to cart

* fixed eslint

* changed compare table

* fixed other category

* changed style for caption

* removed padding

* added comment

* removed other tables

* added styles

* changed tableContentFactory function

* fixed linting errors

* fixed comment from designer

* fixed styles issue

* revert url.ts

* fixed linting errors

* removed redundancy

* resolved conflict

+++ src/components/modules/Compare/CompareModal/TireWithInfoList/TireWithInfoList.tsx
@@ -1,4 +1,5 @@
 import { SiteCatalogProductItem } from '~/data/models//SiteCatalogProductItem';
+import { CSSStylesProp } from '~/lib/constants';
 
 import { NUMBER_OF_TIRES } from '../../Compare.constants';
 import AddTire from '../../ProductToCompare/AddTire';
@@ -6,10 +7,11 @@ import TireWithInfo from './TireWithInfo';
 import styles from './TireWithInfoList.styles';
 
 interface Props {
-  onAddTire: () => void;
-  onRemove: (productId: string) => void;
+  customRootStyle?: CSSStylesProp;
+  onAddTire?: () => void;
+  onRemove?: (productId: string) => void;
   productList: SiteCatalogProductItem[];
-  setRemovingProductIndex: (index: number) => void;
+  setRemovingProductIndex?: (index: number) => void;
 }
 
 function TireWithInfoList({
@@ -17,23 +19,24 @@ function TireWithInfoList({
   onRemove,
   setRemovingProductIndex,
   onAddTire,
+  customRootStyle,
 }: Props) {
   const onClose = (productId: string) => {
-    onRemove(productId);
+    onRemove && onRemove(productId);
   };
 
   return (
-    <div css={[styles.root, styles.background]}>
+    <div css={[styles.root, styles.background, customRootStyle]}>
       {productList.map((product, index) => (
         <TireWithInfo
           key={`tire-with-info-${product.productId}-${product.size}-${index}`}
           index={index}
           product={product}
-          onClose={onClose}
+          onClose={onRemove && onClose}
           setRemovingProductIndex={setRemovingProductIndex}
         />
       ))}
-      {productList.length < NUMBER_OF_TIRES.MAX && (
+      {onAddTire && productList.length < NUMBER_OF_TIRES.MAX && (
         <div css={styles.addTire}>
           <AddTire onAddTire={onAddTire} isBig hasBackground />
         </div>
